Gather all network config
