--- packages/ui/src/app/config.ts
@@ -1,7 +1,0 @@
-export const OLYMPIA_TESTNET_NODE_SOCKET =
-  process.env.REACT_APP_OLYMPIA_TESTNET_NODE_SOCKET || 'wss://olympia-dev.joystream.app/rpc'
-export const OLYMPIA_TESTNET_QUERY_NODE =
-  process.env.REACT_APP_OLYMPIA_TESTNET_QUERY_NODE || 'https://olympia-dev.joystream.app/query/server/graphql'
-export const OLYMPIA_TESTNET_QUERY_NODE_SOCKET =
-  process.env.REACT_APP_OLYMPIA_TESTNET_QUERY_NODE_SOCKET || 'wss://olympia-dev.joystream.app/query/server/graphql'
-export const MEMBERSHIP_FAUCET_URL = process.env.REACT_APP_MEMBERSHIP_FAUCET_URL || 'http://localhost:4000/register'
