+++ packages/ui/src/app/config/index.ts
@@ -1,0 +1,1 @@
+export * from './network'
