+++ packages/ui/src/app/index.ts
@@ -1,2 +1,3 @@
 export * from './App'
 export { WaitForAPI } from '@/app/components/WaitForAPI'
+export { MEMBERSHIP_FAUCET_ENDPOINT } from '@/app/config'
