+++ packages/ui/src/app/pages/Settings/Settings.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 
 import { PageLayout, PageHeaderWrapper } from '@/app/components/PageLayout'
+import { NetworkType } from '@/app/config'
 import { MainPanel } from '@/common/components/page/PageContent'
 import { PageTitle } from '@/common/components/page/PageTitle'
 import { SimpleSelect } from '@/common/components/selects'
-import { NetworkType, useNetwork } from '@/common/hooks/useNetwork'
+import { useNetwork } from '@/common/hooks/useNetwork'
 
 export const Settings = () => {
   const options: NetworkType[] = ['local', 'local-mocks', 'olympia-testnet']
