+++ packages/ui/src/common/hooks/useNetwork.ts
@@ -1,10 +1,10 @@
-import { useLocalStorage } from './useLocalStorage'
+import { NetworkType } from '@/app/config'
 
-export type NetworkType = 'local' | 'local-mocks' | 'olympia-testnet'
+import { useLocalStorage } from './useLocalStorage'
 
 export const useNetwork = () => {
   const [network, setNetwork] = useLocalStorage<NetworkType>('network')
-  const resolvedNetwork = network || 'local'
+  const resolvedNetwork = network ?? 'local'
 
   return [resolvedNetwork, setNetwork] as const
 }
